+++ .eslintrc.js
@@ -2,6 +2,6 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 module.exports = {
 	extends: [
-		'@nextcloud',
+		'@nextcloud/eslint-config/vue3',
 	]
 }